requirements.txt
flask
torch
transformers
pillow
rasterio
geopandas
pandas
scikit-image
shapely
